% Auto-generated: do not edit by hand
\name{''DashQrGenerator}

\alias{''DashQrGenerator}

\title{DashQrGenerator component}

\description{
Dash QR Code Generator for more information: https://www.npmjs.com/package/react-google-qrcode
}

\usage{
''DashQrGenerator(id=NULL, data=NULL, framed=NULL, size=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Character. The data to encode.}

\item{framed}{Logical. Adds a frame to the image}

\item{size}{Numeric. Image size in pixels (width x height). Min value: 30, Max value: 547}

\item{style}{Named list. The style of the QR code.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

