% Auto-generated: do not edit by hand
\name{dashQrReader}

\alias{dashQrReader}

\title{DashQrReader component}

\description{
Dash QR Code Reader for more information: https://www.npmjs.com/package/react-qr-reader
}

\usage{
dashQrReader(id=NULL, className=NULL, constraints=NULL,
containerStyle=NULL, result=NULL, scanDelay=NULL,
style=NULL, videoContainerStyle=NULL, videoId=NULL,
videoStyle=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. ClassName for the container element.}

\item{constraints}{Named list. Specify which camera should be used (if available)}

\item{containerStyle}{Named list. Style object for the container element.}

\item{result}{Character. decoded data}

\item{scanDelay}{Numeric. The scan period for the QR hook}

\item{style}{Named list. The style of the QR code.}

\item{videoContainerStyle}{Named list. Style object for the video container element.}

\item{videoId}{Character. The ID for the video element}

\item{videoStyle}{Named list. Style object for the video element.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

